+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Dev
+- Fix minify-css-idents/css-loader generating too much idents
+
 # Version 1.0.2
 - Fix minify-css-idents/css-loader breaking options for other loaders